requirements.txt
cryptography
dnacentersdk
GitPython
nornir
nornir_netmiko
PyYAML
requests
pynetbox